Changed "cpstate->pstate." to "pstate->"
There are some sections of the code that uses "cpstate->pstate." and "pstate->" at others, but they are the same thing. Changed all to "pstate->" so that the code is more readable.
Furthermore, the "pstate" variable was being created as a downcast of the "cpstate". Since there is a "pstate" member for the "cpstate", giving the variable the value of "&cpstate->pstate" seems better.