WIP for varnullingrels issue
This is a work in progress for the new field that was
added to the struct Var called "varnullingrels".
According to the documentation, this field is responsible
for marking the Vars as nullable, if they are coming from
a JOIN, either LEFT JOIN, RIGHT JOIN, or FULL OUTER JOIN.
The changes were made following an "optional match" clause
which is being treated as a LEFT JOIN from our extension.

A function markRelsAsNulledBy is added because its internal
in Postgres and doesn't belong in a header file, therefore
it can't be exported. This function is added before the creation
of the Vars from the "make_vertex_expr" and "make_edge_expr",
to correctly mark the specific PNSI as nullable, so later in
the planner stage, the Vars will be correctly nulled.

Signed-off-by: Panagiotis Foliadis <pfoliadis@hotmail.com>